Simplify navigation setup
