+++ src/app.d.ts
@@ -7,7 +7,3 @@ declare namespace App {
 	// interface Error {}
 	// interface Platform {}
 }
-
-declare module '@fortawesome/pro-solid-svg-icons/index.es' {
-	export * from '@fortawesome/pro-solid-svg-icons';
-}
