+++ src/modules.d.ts
@@ -1,0 +1,3 @@
+declare module '@fortawesome/pro-solid-svg-icons/index.es' {
+	export * from '@fortawesome/pro-solid-svg-icons';
+}
