--- src/types/NavigationItem.ts
@@ -1,4 +1,0 @@
-export interface NavigationItem {
-	title: string;
-	href: string;
-}
